Connect with local computer command line interface
Developed the command line interface with a local computer for weather data. This weather data has contained, more than 10 columns and mess up data so, This code helps to extract four columns out of that. So, the user can able to connect the data with a local computer and can able to extract some columns